minor changes to DB/migrations

+++ src/main/kotlin/com/scrumdapp/groupservice/dto/GroupUserDto.kt
@@ -1,6 +1,6 @@
 package com.scrumdapp.groupservice.dto
 
 data class GroupUserDto(
-    val userId: Int,
-    val groupId: Int
+    val userId: Long,
+    val groupId: Long
 )
